+++ vscode/src/core/fetchers/JsrFetcher.ts
@@ -1,7 +1,6 @@
 
 import { versions } from "../../api/indexes/jsr";
 import { Settings } from "../../config";
-import compareVersions from "../../semver/compareVersions";
 import { fetcherCatch } from "../../utils/errors";
 import Dependency from "../Dependency";
 import { Fetcher } from "./fetcher";
@@ -13,9 +12,11 @@ export class JsrFetcher extends Fetcher {
     return async function (dep: Dependency): Promise<Dependency> {
       const checkVersion = versions(dep.item.key, dep.item.value);
       return checkVersion.then((mod) => {
-        const versions = mod.versions
-          .filter((i: string) => i !== "" && i !== undefined && !base.checkUnstables(Settings.npm.unstableFilter, i, dep.item.value!))
-          .sort(compareVersions).reverse();
+        const versions = base.filterAndSortVersions(
+          mod.versions,
+          dep.item.value,
+          Settings.npm.unstableFilter
+        );
         dep.versions = versions;
         dep.item.latestVersion = mod.latestVersion;
         dep.item.value = dep.item.value === "latest" ? mod.latestVersion : dep.item.value;