+++ vscode/src/core/fetchers/NuGetFetcher.ts
@@ -1,6 +1,5 @@
 import { versions } from "../../api/indexes/nuget";
 import { Settings } from "../../config";
-import compareVersions from "../../semver/compareVersions";
 import { fetcherCatch } from "../../utils/errors";
 import Dependency from "../Dependency";
 import { Fetcher } from "./fetcher";
@@ -14,12 +13,11 @@ export class NuGetFetcher extends Fetcher {
       }
       return versions(dep.item.key)
         .then((nugetPackage) => {
-          const versions = nugetPackage.versions
-            .filter((i: string) => i !== "" && i !== undefined && !base.checkUnstables(Settings.csharp.unstableFilter, i, dep.item.value!))
-            .sort(compareVersions)
-            .reverse();
-
-          dep.versions = versions;
+          dep.versions = base.filterAndSortVersions(
+            nugetPackage.versions,
+            dep.item.value,
+            Settings.csharp.unstableFilter
+          );
           return dep;
         })
         .catch(fetcherCatch(dep));