+++ vscode/src/core/fetchers/PackagistFetcher.ts
@@ -1,7 +1,6 @@
 
 import { versions } from "../../api/indexes/packagist";
 import { Settings } from "../../config";
-import compareVersions from "../../semver/compareVersions";
 import { fetcherCatch } from "../../utils/errors";
 import Dependency from "../Dependency";
 import { Fetcher } from "./fetcher";
@@ -13,10 +12,11 @@ export class PackagistFetcher extends Fetcher {
     return async function (dep: Dependency): Promise<Dependency> {
       const checkVersion = isLatest ? versions(dep.item.key) : versions(dep.item.key);
       return checkVersion.then((mod: any) => {
-        const versions = mod.versions
-          .filter((i: string) => i !== "" && i !== undefined && !base.checkUnstables(Settings.php.unstableFilter, i, dep.item.value!))
-          .sort(compareVersions)
-          .reverse();
+        const versions = base.filterAndSortVersions(
+          mod.versions,
+          dep.item.value,
+          Settings.php.unstableFilter
+        );
         dep.versions = versions;
         return dep;
       }).catch(fetcherCatch(dep));