+++ vscode/src/core/fetchers/TerraformFetcher.ts
@@ -1,6 +1,5 @@
 import { versions } from "../../api/indexes/terraform";
 import { Settings } from "../../config";
-import compareVersions from "../../semver/compareVersions";
 import { fetcherCatch } from "../../utils/errors";
 import Dependency from "../Dependency";
 import { Fetcher } from "./fetcher";
@@ -11,9 +10,11 @@ export class TerraformFetcher extends Fetcher {
     const base = this;
     return async function (dep: Dependency): Promise<Dependency> {
       return versions(dep.item.key).then((mod) => {
-        const versionList = mod.versions
-          .filter((i: string) => i !== "" && i !== undefined && !base.checkUnstables(Settings.terraform.unstableFilter, i, dep.item.value!))
-          .sort(compareVersions).reverse();
+        const versionList = base.filterAndSortVersions(
+          mod.versions,
+          dep.item.value,
+          Settings.terraform.unstableFilter
+        );
         dep.versions = versionList;
         return dep;
       }).catch(fetcherCatch(dep));